+++ GameMadness/Engine/Engine.h
@@ -15,7 +15,7 @@ public:
 	static int SCREEN_WIDTH;
 	static int SCREEN_HEIGHT;
 
-	static double GetDT(); //get delta time
+	static float GetDT(); //get delta time
 
 	Engine();
 	~Engine();
@@ -32,8 +32,8 @@ private:
 	
 	//delta time, the time it took for a frame / cycle to complete. we'll multiply this by the speed to enusure everything moves at a constant rate 
 	//regardless of whether you have a faster machine
-	static double dt; 
-	double lastTime; //last time anything happened, in miliseconds
+	static float dt; 
+	float lastTime; //last time anything happened, in miliseconds
 
 };
